Not finished yet, the app does not run. There's a bug somewhere (Unable to instantiate service com.abominableshrine.taptounlock.UnlockService: java.lang.NullPointerException)
