+++ app/src/main/java/com/abominableshrine/taptounlock/SensorListenerService.java
@@ -1,10 +1,8 @@
 package com.abominableshrine.taptounlock;
 
 import android.app.Service;
-import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.content.IntentFilter;
 import android.hardware.Sensor;
 import android.hardware.SensorManager;
 import android.os.IBinder;
@@ -15,9 +13,6 @@ public class SensorListenerService extends Service {
     private static boolean DEBUG;
     private static SensorManager mSensorManager;
     private SensorEventHandler mSensorEventHandler;
-    private static Boolean running; // Boolean class to enable check for null
-    private IntentFilter screenOffFilter;
-    private static ScreenOffBroadcastReceiver mScreenOffBroadcastReceiver;
 
     public SensorListenerService() {
         DEBUG = AppConstants.DEBUG;
@@ -39,13 +34,6 @@ public class SensorListenerService extends Service {
                 mSensorManager.getDefaultSensor(Sensor.TYPE_ACCELEROMETER),
                 SensorManager.SENSOR_DELAY_FASTEST);
 
-        // initialization of the ScreenOffBroadcastReceiver
-        mScreenOffBroadcastReceiver = new ScreenOffBroadcastReceiver();
-        screenOffFilter = new IntentFilter();
-        screenOffFilter.addAction(Intent.ACTION_SCREEN_OFF);
-        registerReceiver(mScreenOffBroadcastReceiver, screenOffFilter);
-
-        running = Boolean.TRUE;
         return START_STICKY;
     }
 
@@ -54,37 +42,9 @@ public class SensorListenerService extends Service {
         if (DEBUG) Log.d(AppConstants.TAG, "SensorListenerService is being destroyed");
         SensorEventHandler.lockDevice();
         mSensorManager.unregisterListener(mSensorEventHandler);
-        unregisterReceiver(mScreenOffBroadcastReceiver);
-        running = Boolean.FALSE;
         super.onDestroy();
     }
 
-    public static boolean isRunning() {
-        return Boolean.TRUE.equals(running);
-    }
 
-    /* This BroadcastReceiver should detect when the screen of the device is turned off. If so,
-     * the device has to be locked again. */
-    public static class ScreenOffBroadcastReceiver extends BroadcastReceiver {
-        @Override
-        public void onReceive(Context context, Intent intent) {
-            if (intent.getAction().equals(Intent.ACTION_SCREEN_OFF)) {
-                if (!SensorListenerService.isRunning()) {
-                    if (DEBUG) Log.d(AppConstants.TAG, "Screen is turned off but the " +
-                            "SensorListenerService is not running. Aborting.");
-                    return;
-                }
-                Log.d(AppConstants.TAG, "Screen turned off. locking device");
-                SensorEventHandler.lockDevice();
-            }
-        }
-    }
 
-    public void stopReceiver() {
-        unregisterReceiver(mScreenOffBroadcastReceiver);
-    }
-
-    public void startReceiver() {
-        registerReceiver(mScreenOffBroadcastReceiver, screenOffFilter);
-    }
 }
