+++ app/src/main/java/com/abominableshrine/taptounlock/UnlockService.java
@@ -1,11 +1,17 @@
 package com.abominableshrine.taptounlock;
 
 import android.app.Service;
+import android.content.BroadcastReceiver;
+import android.content.ComponentName;
+import android.content.Context;
 import android.content.Intent;
-import android.os.Handler;
+import android.content.IntentFilter;
+import android.content.ServiceConnection;
+import android.os.Bundle;
 import android.os.IBinder;
 import android.os.Message;
 import android.os.Messenger;
+import android.os.RemoteException;
 import android.util.Log;
 import android.view.KeyEvent;
 
@@ -19,35 +25,73 @@ import java.io.IOException;
  */
 public class UnlockService extends Service {
     private final static Boolean DEBUG = AppConstants.DEBUG;
-    private Messenger mMessenger = new Messenger (new MessageHandler());
+    private static UnlockService instance;
+
+    private boolean mBound;
+    private static Boolean running; // Boolean class to enable check for null
+    private Messenger mUnlockServiceMessenger = null;
+    private static ScreenOffBroadcastReceiver mScreenOffBroadcastReceiver;
+    private IntentFilter screenOffFilter;
+    private final Intent tapPatternDetectorServiceIntent = new Intent(this, TapPatternDetectorService.class);
+    private ServiceConnection mUnlockServiceConnection = new ServiceConnection() {
+        public void onServiceConnected(ComponentName className, IBinder service) {
+            mUnlockServiceMessenger = new Messenger(service);
+            mBound = true;
+        }
+
+        public void onServiceDisconnected(ComponentName className) {
+            mUnlockServiceMessenger = null;
+            mBound = false;
+        }
+    };
 
     protected static final int MSG_UNLOCK = 1;
     protected static final int MSG_LOCK = 2;
 
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
+        instance = this;
+        running = true;
+        bindService(tapPatternDetectorServiceIntent, mUnlockServiceConnection, Context.BIND_AUTO_CREATE);
+
+        // initialization of the ScreenOffBroadcastReceiver
+        mScreenOffBroadcastReceiver = new ScreenOffBroadcastReceiver();
+        screenOffFilter = new IntentFilter();
+        screenOffFilter.addAction(Intent.ACTION_SCREEN_OFF);
+        registerReceiver(mScreenOffBroadcastReceiver, screenOffFilter);
+
+        // Tell the TapPatternDetectorService what pattern to search for
+        Message mMessage = Message.obtain(null, MSG_UNLOCK);
+        Bundle messageData = new Bundle();
+        messageData.putString("UnlockPattern", "This could be any parcelable data type, a string or even a byte[]");
+        mMessage.setData(messageData);
+        try {
+            mUnlockServiceMessenger.send(mMessage);
+        } catch (RemoteException e) {
+            e.printStackTrace();
+        }
+
         return START_STICKY;
     }
 
     @Override
     public IBinder onBind(Intent intent) {
-        return mMessenger.getBinder();
+        throw new UnsupportedOperationException("UnlockService.onBind(Intent intent) not implemented!");
     }
 
-    class MessageHandler extends Handler {
-        @Override
-        public void handleMessage(Message msg) {
-            switch (msg.what) {
-                case MSG_UNLOCK:
-                    unlock();
-                    break;
-                case MSG_LOCK:
-                    lock();
-                    break;
-                default:
-                    super.handleMessage(msg);
-            }
+    @Override
+    public void onDestroy() {
+        unregisterReceiver(mScreenOffBroadcastReceiver);
+        lock();
+
+        // Unbind from the service
+        if (mBound) {
+            unbindService(mUnlockServiceConnection);
+            mBound = false;
         }
+
+        running = false;
+        super.onDestroy();
     }
 
     /**
@@ -56,6 +100,7 @@ public class UnlockService extends Service {
      * @see #unlock()
      */
     private static void lock() {
+
         String[] shellCommands = {
                 "cd /data/system",
                 "mv passwordtemp.key password.key",
@@ -74,6 +119,11 @@ public class UnlockService extends Service {
      * @see #lock()
      */
     private static void unlock() {
+        // When the device is being unlocked, the ScreenOffBroadcastReceiver should first be
+        // stopped so it does not disturb the unlocking process. Afterwards, we turn it on
+        // again.
+        instance.stopReceiver();
+
         String[] shellCommands = {
                 "cd /data/system",
                 "mv password.key passwordtemp.key",
@@ -82,6 +132,9 @@ public class UnlockService extends Service {
         };
         runAsRoot(shellCommands);
         if (DEBUG) Log.d(AppConstants.TAG, "Device unlocked");
+
+        // TODO: Test if this is the right place to start the receiver again.
+        instance.startReceiver();
     }
 
     /**
@@ -105,4 +158,34 @@ public class UnlockService extends Service {
             e.printStackTrace();
         }
     }
+
+    /* This BroadcastReceiver should detect when the screen of the device is turned off. If so,
+ * the device has to be locked again. */
+    public static class ScreenOffBroadcastReceiver extends BroadcastReceiver {
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            if (intent.getAction().equals(Intent.ACTION_SCREEN_OFF)) {
+                if (!isRunning()) {
+                    if (DEBUG) Log.d(AppConstants.TAG, "Screen is turned off but the " +
+                            "UnlockService is not running. Aborting.");
+                    return;
+                }
+                Log.d(AppConstants.TAG, "Screen turned off. locking device");
+                SensorEventHandler.lockDevice();
+            }
+        }
+    }
+
+    public void stopReceiver() {
+        unregisterReceiver(mScreenOffBroadcastReceiver);
+    }
+
+    public void startReceiver() {
+        registerReceiver(mScreenOffBroadcastReceiver, screenOffFilter);
+    }
+
+    public static boolean isRunning() {
+        return Boolean.TRUE.equals(running);
+    }
+
 }
